A0-2985: Move functions to utils
