+++ packages/page-addresses/src/util.tsx
@@ -1,13 +1,17 @@
 // Copyright 2017-2023 @polkadot/app-addresses authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import type { ApiPromise } from '@polkadot/api';
 import type { KeyringAddress } from '@polkadot/ui-keyring/types';
 import type { SortedAccount } from './types.js';
 
+import { resolveDomainToAddress } from '@azns/resolver-core';
 import React from 'react';
 
-import { Menu } from '@polkadot/react-components';
+import { Menu, systemNameToChainId } from '@polkadot/react-components';
 import { keyring } from '@polkadot/ui-keyring';
+import { hexToU8a } from '@polkadot/util';
+import { ethereumEncode } from '@polkadot/util-crypto';
 
 export function createMenuGroup (items: (React.ReactNode | false | undefined | null)[]): React.ReactNode | null {
   const filtered = items.filter((item): item is React.ReactNode => !!item);
@@ -62,4 +66,34 @@ export function sortAccounts (addresses: string[], favorites: string[]): SortedA
           ? 1
           : -1
     );
+}
+
+export function getValidatedAddress (address: string, isEthereum: boolean): string | undefined {
+  try {
+    if (isEthereum) {
+      const rawAddress = hexToU8a(address);
+
+      return ethereumEncode(rawAddress);
+    }
+
+    const publicKey = keyring.decodeAddress(address);
+
+    return keyring.encodeAddress(publicKey);
+  } catch {
+    return undefined;
+  }
+}
+
+export async function getAddressFromDomain (domain: string, { api, systemChain }: {api: ApiPromise, systemChain: string}): Promise<string | null | undefined> {
+  const chainId = systemNameToChainId.get(systemChain);
+
+  if (!chainId) {
+    return;
+  }
+
+  try {
+    return (await resolveDomainToAddress(domain, { chainId, customApi: api })).address;
+  } catch {
+    return undefined;
+  }
 }
